may 24 class
