+++ live_python_code/June_2022/calculator_day8.py
@@ -1,0 +1,14 @@
+# 26.6.2022
+# sunday
+
+from python_module_demo_day8 import add_two_numbers, sub_two_numbers
+
+print("addition of 23, 33 : ", add_two_numbers(23, 33))
+# addition of 23, 33 :  56
+
+print("subtraction of 23, 33 : ", sub_two_numbers(23, 33))
+# subtraction of 23, 33 : -10
+
+print("sai ram")
+
+# LEGB Rule
