+++ live_python_code/June_2022/python_day_5_doubts.py
@@ -1,0 +1,44 @@
+first_o_index = 'CARTOONNNOO'.index('O')
+print(first_o_index)
+second_o_index = 'CARTOON'.index('O', first_o_index + 1)
+third_o_index = 'CARTOON'.index('O', first_o_index + 2)
+print(third_o_index)
+# fourth_o_index = 'CARTOON'.index('O', third_o_index + 1)
+
+# import sys
+
+# sys.exit(0)
+# # break
+# # assert
+
+# i = 1
+# while (i < 2):
+# print(i)
+
+# stringg = "ramesh"
+# len(stringg)
+# print(id(print("Sai ram")))
+
+# num = ['h','e','l','l','o',1,2,3, [1,2,3], (1,2,3) ]
+
+# "test data"
+
+# numbers = [0,1,2,3,4,5,6,7,8,9]
+
+# print(numbers[::-1])
+
+# f - strings
+# only python 3
+# print(f"the sum of {a} and {b} is {sum}")
+
+# # older versions till python 2
+# print("the sum of %d and %d is %d".format(a,b,sum))
+
+#         *
+#      *python*
+#    *is  *  a    *
+#  * good  * programming * language *
+# * to * learn * for * beginners *
+
+# variable = [10, 20, 30, 40]
+# len(variable)
