+++ live_python_code/June_2022/python_module_demo_day8.py
@@ -1,0 +1,8 @@
+def add_two_numbers(number1, number2):
+    return number1 + number2
+
+def sub_two_numbers(number1, number2):
+    return number1 - number2
+
+def mul_two_numbers(number1, number2):
+    return number1 * number2
