--- live_python_code/python_module_demo_day8.py
@@ -1,8 +1,0 @@
-def add_two_numbers(number1, number2):
-    return number1 + number2
-
-def sub_two_numbers(number1, number2):
-    return number1 - number2
-
-def mul_two_numbers(number1, number2):
-    return number1 * number2
